+++ pennylane/compiler/python_compiler/transforms/quantum/cancel_inverses.py
@@ -20,8 +20,8 @@ from xdsl import context, passes, pattern_rewriter
 from xdsl.dialects import builtin, func
 from xdsl.ir import Operation
 
-from ..dialects.quantum import CustomOp
-from .api import compiler_transform
+from ...dialects.quantum import CustomOp
+from ...pass_api import compiler_transform
 
 self_inverses = [
     "Identity",
