+++ tests/python_compiler/transforms/quantum/test_xdsl_combine_global_phases.py
