+++ tests/python_compiler/transforms/quantum/test_xdsl_merge_rotations.py
